requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Configuration & YAML Processing
PyYAML>=6.0
pydantic-settings>=2.0.0
setuptools>=65.0.0

# Machine Learning & NLP
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
spacy>=3.7.0

# Vector Search & Similarity
faiss-cpu>=1.7.4
qdrant-client>=1.6.0

# Document Processing
PyMuPDF>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-magic>=0.4.27

# Data Processing & Export
pandas>=2.0.0
aiofiles>=23.0.0
scikit-learn>=1.3.0

# HTTP & Async Operations
aiohttp>=3.9.0
requests>=2.31.0

# Database & Storage
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Security & Encryption
cryptography>=41.0.0
PyJWT>=2.8.0

# Advanced Features (Phase 10)
boto3>=1.34.0  # For S3 backup
paramiko>=3.3.0  # For SFTP backup
imagehash>=4.3.1  # For image duplicate detection

# System Monitoring
psutil>=5.9.0

# AI/LLM Integration
openai>=1.3.0

# External Service Integrations (Phase 10 Advanced Features)
google-api-python-client>=2.100.0  # For Google Drive integration
google-auth>=2.22.0  # For Google authentication
google-auth-oauthlib>=1.1.0  # For OAuth flows
atlassian-python-api>=3.41.0  # For Confluence integration

# Additional dependencies for production
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
celery>=5.3.0  # For background task processing

# Development and Testing (install with: pip install -r requirements.txt[dev])
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.3.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
bandit>=1.7.5
safety>=2.3.0
pre-commit>=3.5.0